Create README.md
**_Bike Sales Analysis Dashboard_**

**Problem Statement**

In the dynamic landscape of the biking industry, understanding the factors influencing sales revenue is crucial for strategic decision-making. This Excel project aims to analyze biker sales revenue based on various demographic and regional factors, such as age, age brackets, income, and region.

**Dataset Used**

A historical dataset of demographics of bike buyers  with mutliple rows and columns spanning. Data is stored in one table and the analysis in present in the rest 3 tables.